Support deserialisation of messages
`IQueueMessage` instances store the contents of the message in the
`Body` property with a type of `byte[]`. The `EnqueueAsync` API allowed
callers to serialise arbitrary types to an `IQueueMessage`, but no
methods were provided to perform the reverse process: callers would
need to figure out the correct deserialisation themselves.

This change introduces a `GetBody<T>()` extension method on the
`IQueueMessage` interface that performs the correct deserialisation.