Fix bug in the disposal of in memory queue clients
Previously, disposing of an instance of `InMemoryQueueClient` would
clear the internal queue of all messages. This could cause unexpected
behaviour, since other instances of this class that share the same queue
name would suddenly see all messages removed from the queue.

This change introduces a counter of the number of queue instances that
share the same queue name. The internal queue of messages for that queue
name is now only cleared once this counter reaches zero.