Remove invalid in-memory client test
We previously had a test that tested that messages can be dequeued
across disposed queue client istances. Since we disposed of queues one
all clients referencing that queue (for the in-memory implementation),
the scenario under test was never actually supported. We'd expect that
this implementation is used in tests, where a test fixture would be
used to maintain a reference to the queue for the lifetime of the tests.