Support completing renewed Azure Storage Queue messages
Previously, we were not updaing the pop receipt for Azure Storage Queue
messages when we renewed the lease on the message. This caused calling
`AzureStorageQueueClient.CompleteAsync()` to throw since we were still
using the old pop receipt. This change ensures that we update this
receipt each time that we renew a message.