Fix Service Bus socket exhaustion (#28)
This change introduces caching and sharing of Service Bus connections
across queue client instances. Previously, a new connection to Service
Bus was created for each instance of the `AzureServiceBusQueueClient`
created. This could cause socket exceptions if a large number of these
are created at the same time, as we've observed during load testing.

With this change, a new connection is created once per unique
connection string and queue name, and this is disposed only once all
connections to the queue have been released.

Resolves #20.