Remove casing of Service Bus variables
The AzureServiceBusQueueClient converted the connection string and queue name to their lowercase forms when creating the key for the queues dictionairy. However, the two lowercased variables were then being used to initialise the Service Bus queue, causing an unauthorised exception to be thrown.

The casing has been removed, which is inline with how Azure WebJobs interacts with Service Bus.